requirements.txt
pyyaml>=6.0
requests>=2.31.0
anthropic>=0.40.0
openai>=1.0.0
google-generativeai>=0.3.0
